requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
# Add project-specific packages here, e.g.:
# openai
# pandas
openai
pandas
